# Introduce `UPSERT` in `set_trial_system_attr`