@c-bata 2024-11-01
@gen740 @nabenabe0928 Could you review this PR if you have time?